Move TmpTD sprite loading into Mods.Common.
