+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -116,6 +116,7 @@
     <Compile Include="SpriteLoaders\ShpTDLoader.cs" />
     <Compile Include="SpriteLoaders\ShpTSLoader.cs" />
     <Compile Include="SpriteLoaders\TmpRALoader.cs" />
+    <Compile Include="SpriteLoaders\TmpTDLoader.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
