+++ mods/cnc/mod.yaml
@@ -211,4 +211,4 @@ Missions:
 
 SupportsMapsFrom: cnc
 
-SpriteFormats: ShpTD, ShpTS, TmpRA
+SpriteFormats: ShpTD, TmpTD, ShpTS, TmpRA
