+++ mods/ra/mod.yaml
@@ -208,4 +208,4 @@ Missions:
 
 SupportsMapsFrom: ra
 
-SpriteFormats: ShpTD, TmpRA, ShpTS
+SpriteFormats: ShpTD, TmpRA, TmpTD, ShpTS
